\name{PIAAC-package}
\docType{package}
\title{
Programme for the International Assessment of Adult Competencies (PIAAC)
}
\description{
This is an R package that aggregates data from PIAAC 2013.

There are 22 datasets in this package, each with data from single country. Information from the PIAAC website:

The data collection for the Survey of Adult Skills (PIAAC) took place from 1 August 2011 to 31 March 2012 in most participating countries. Around 166 000 adults, representing 724 million adults aged 16 to 65, were surveyed in 24 countries and sub-national regions in the official language/s of the countries

More information about data, coodebook and analysis manual can be found here: http://www.oecd.org/site/piaac/publicdataandanalysis.htm}
\details{
\tabular{ll}{
Package: \tab PIAAC\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-24\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Przemyslaw Biecek (OECD)

Maintainer: Przemyslaw Biecek <Przemyslaw.Biecek@gmail.com>
}
\keyword{package}
\seealso{
}
\examples{
}
